refactor: move to go-project-layout
