--- frontend/.env
@@ -1,1 +1,0 @@
-BACKEND_PORT=9000
