+++ internal/models/sql/tag.sql
