+++ web/src/components/shared/ErrorBoundary/index.tsx
